♻️ moved transitioonEnd function to globals/helper
Can be used if necessary in other components
